@@ -5,7 +5,7 @@ homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
   flutter: ">=2.5.0"
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,7 +24,7 @@ dev_dependencies:
     sdk: flutter
   golden_toolkit: ^0.12.0
   lint: any
-  mocktail: ^0.2.0
+  mocktail: ^0.3.0
 
 flutter:
   assets:
