@@ -64,9 +64,11 @@ class _PostScreen extends StatelessWidget {
           actions: [
             IconButton(
               icon: const Icon(Icons.open_in_browser),
-              onPressed: () => launch(post.link),
+              onPressed: () => launchUrl(Uri.parse(post.link)),
             ),
-            const PopupMenu(),
+            const PopupMenu(
+              toggleIsSelectable: true,
+            ),
           ],
         ),
         body: Padding(
@@ -85,6 +87,7 @@ class _PostScreen extends StatelessWidget {
                 const SizedBox(height: 8),
                 HtmlWidget(
                   post.content,
+                  isSelectable: context.isSelectable,
                 ),
               ],
             ),
