@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:demo_app/screens/golden.dart';
+import 'package:demo_app/widgets/popup_menu.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -9,9 +10,23 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 
-import '../../packages/fwfh_chewie/test/_.dart';
+import '../../packages/fwfh_chewie/test/mock_video_player_platform.dart';
+import '../../packages/fwfh_webview/test/mock_webview_platform.dart';
 
-void _test(String name, String html) => testGoldens(
+final goldenSkipEnvVar = Platform.environment['GOLDEN_SKIP'];
+final goldenSkip = goldenSkipEnvVar == ''
+    ? Platform.isLinux
+        ? null
+        : 'Linux only'
+    : 'GOLDEN_SKIP=$goldenSkipEnvVar';
+
+void _test(
+  String name,
+  String html, {
+  bool isSelectable = false,
+  double textScaleSize = 1.0,
+}) =>
+    testGoldens(
       name,
       (tester) async {
         const platform = TargetPlatform.android;
@@ -20,60 +35,54 @@ void _test(String name, String html) => testGoldens(
         final key = UniqueKey();
 
         await tester.pumpWidgetBuilder(
-          Golden(name, html, targetKey: key),
+          PopupMenuStateProvider(
+            builder: (_) => Golden(name, html, targetKey: key),
+            initialIsSelectable: isSelectable,
+          ),
           wrapper: materialAppWrapper(
             platform: platform,
             theme: ThemeData.light(),
           ),
           surfaceSize: const Size(400, 1200),
+          textScaleSize: textScaleSize,
         );
 
         await screenMatchesGolden(tester, name, finder: find.byKey(key));
         debugDefaultTargetPlatformOverride = null;
         WidgetFactory.debugDeterministicLoadingWidget = false;
       },
-      skip: !Platform.isLinux,
+      skip: goldenSkip != null,
     );
 
 void main() {
   mockVideoPlayerPlatform();
+  mockWebViewPlatform();
 
   const audioSessionMc = MethodChannel('com.ryanheise.audio_session');
   audioSessionMc.setMockMethodCallHandler((_) async {});
 
-  const platformViewsMc = MethodChannel('flutter/platform_views');
-  platformViewsMc.setMockMethodCallHandler((_) async {});
-
-  for (final name in [
-    'dev.flutter.pigeon.DownloadListenerHostApi.create',
-    'dev.flutter.pigeon.WebChromeClientHostApi.create',
-    'dev.flutter.pigeon.WebSettingsHostApi.create',
-    'dev.flutter.pigeon.WebSettingsHostApi.setBuiltInZoomControls',
-    'dev.flutter.pigeon.WebSettingsHostApi.setDisplayZoomControls',
-    'dev.flutter.pigeon.WebSettingsHostApi.setDomStorageEnabled',
-    'dev.flutter.pigeon.WebSettingsHostApi.setJavaScriptCanOpenWindowsAutomatically',
-    'dev.flutter.pigeon.WebSettingsHostApi.setJavaScriptEnabled',
-    'dev.flutter.pigeon.WebSettingsHostApi.setLoadWithOverviewMode',
-    'dev.flutter.pigeon.WebSettingsHostApi.setMediaPlaybackRequiresUserGesture',
-    'dev.flutter.pigeon.WebSettingsHostApi.setSupportMultipleWindows',
-    'dev.flutter.pigeon.WebSettingsHostApi.setSupportZoom',
-    'dev.flutter.pigeon.WebSettingsHostApi.setUseWideViewPort',
-    'dev.flutter.pigeon.WebSettingsHostApi.setUserAgentString',
-    'dev.flutter.pigeon.WebViewClientHostApi.create',
-    'dev.flutter.pigeon.WebViewHostApi.create',
-    'dev.flutter.pigeon.WebViewHostApi.loadUrl',
-    'dev.flutter.pigeon.WebViewHostApi.setDownloadListener',
-    'dev.flutter.pigeon.WebViewHostApi.setWebChromeClient',
-    'dev.flutter.pigeon.WebViewHostApi.setWebContentsDebuggingEnabled',
-    'dev.flutter.pigeon.WebViewHostApi.setWebViewClient',
-  ]) {
-    BasicMessageChannel(name, const StandardMessageCodec())
-        .setMockMessageHandler((input) async => {'foo': 'bar'});
-  }
-
   final json = File('test/goldens.json').readAsStringSync();
   final map = jsonDecode(json) as Map<String, dynamic>;
   for (final entry in map.entries) {
-    _test(entry.key, entry.value as String);
+    final name = entry.key;
+    final html = entry.value as String;
+    _test(name, html);
+
+    if (name == 'FONT') {
+      _test('x2/$name', html, textScaleSize: 2.0);
+    }
+
+    if (!name.contains('/')) {
+      _test('selectable/$name', html, isSelectable: true);
+
+      if (name == 'FONT') {
+        _test(
+          'selectable/x2/$name',
+          html,
+          isSelectable: true,
+          textScaleSize: 2.0,
+        );
+      }
+    }
   }
 }
