@@ -229,7 +229,13 @@ Future<void> main() async {
       expect(urls, equals(const [href]));
     });
 
-    final goldenSkip = Platform.isLinux ? null : 'Linux only';
+    final goldenSkipEnvVar = Platform.environment['GOLDEN_SKIP'];
+    final goldenSkip = goldenSkipEnvVar == ''
+        ? Platform.isLinux
+            ? null
+            : 'Linux only'
+        : 'GOLDEN_SKIP=$goldenSkipEnvVar';
+
     GoldenToolkit.runWithConfiguration(
       () {
         testGoldens(
@@ -240,7 +246,8 @@ Future<void> main() async {
                 body: Padding(
                   padding: EdgeInsets.all(8.0),
                   child: HtmlWidget(
-                    '<div style="background: black; color: white; width: 200px; height: 200px">'
+                    '<div style="background: black; color: white; '
+                    'width: 200px; height: 200px">'
                     '<ruby>Foo <rt>bar</rt></ruby>'
                     '<div>',
                   ),
