@@ -9,7 +9,7 @@ import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 import 'package:flutter_widget_from_html_core/src/internal/tsh_widget.dart';
 import 'package:html/dom.dart' as dom;
 
-import '../../fwfh_url_launcher/test/_.dart' as fwfh_url_launcher;
+import '../../fwfh_url_launcher/test/mock_url_launcher_platform.dart';
 import '_.dart' as helper;
 
 Future<String> explain(WidgetTester t, HtmlWidget hw) =>
@@ -328,6 +328,33 @@ void main() {
     });
   });
 
+  group('isSelectable', () {
+    const html = 'Foo';
+
+    testWidgets('renders RichText', (WidgetTester tester) async {
+      final hw = HtmlWidget(html, key: helper.hwKey);
+      final explained = await explain(tester, hw);
+      expect(explained, equals('[RichText:(:Foo)]'));
+    });
+
+    testWidgets('renders SelectableText', (WidgetTester tester) async {
+      final hw = HtmlWidget(html, isSelectable: true, key: helper.hwKey);
+      final explained = await explain(tester, hw);
+      expect(explained, equals('[SelectableText:(:Foo)]'));
+    });
+
+    testWidgets('renders onSelectionChanged', (WidgetTester tester) async {
+      final hw = HtmlWidget(
+        html,
+        isSelectable: true,
+        key: helper.hwKey,
+        onSelectionChanged: (_, __) {},
+      );
+      final explained = await explain(tester, hw);
+      expect(explained, equals('[SelectableText:+onSelectionChanged,(:Foo)]'));
+    });
+  });
+
   group('onErrorBuilder', () {
     Future<String?> explain(
       WidgetTester tester, {
@@ -413,8 +440,7 @@ void main() {
   });
 
   group('onTapUrl', () {
-    setUp(fwfh_url_launcher.mockSetup);
-    tearDown(fwfh_url_launcher.mockTearDown);
+    setUp(mockUrlLauncherPlatform);
 
     testWidgets('triggers callback (returns false)', (tester) async {
       const href = 'returns-false';
@@ -434,7 +460,7 @@ void main() {
       expect(await helper.tapText(tester, 'Tap me'), equals(1));
 
       expect(urls, equals(const [href]));
-      expect(fwfh_url_launcher.mockGetLaunchUrls(), equals(const [href]));
+      expect(mockLaunchedUrls, equals(const [href]));
     });
 
     testWidgets('triggers callback (returns true)', (tester) async {
@@ -455,7 +481,7 @@ void main() {
       expect(await helper.tapText(tester, 'Tap me'), equals(1));
 
       expect(urls, equals(const [href]));
-      expect(fwfh_url_launcher.mockGetLaunchUrls(), equals(const []));
+      expect(mockLaunchedUrls, equals(const []));
     });
 
     testWidgets('triggers callback (async false)', (tester) async {
@@ -476,7 +502,7 @@ void main() {
       expect(await helper.tapText(tester, 'Tap me'), equals(1));
 
       expect(urls, equals(const [href]));
-      expect(fwfh_url_launcher.mockGetLaunchUrls(), equals(const [href]));
+      expect(mockLaunchedUrls, equals(const [href]));
     });
 
     testWidgets('triggers callback (async true)', (tester) async {
@@ -497,7 +523,7 @@ void main() {
       expect(await helper.tapText(tester, 'Tap me'), equals(1));
 
       expect(urls, equals(const [href]));
-      expect(fwfh_url_launcher.mockGetLaunchUrls(), equals(const []));
+      expect(mockLaunchedUrls, equals(const []));
     });
 
     testWidgets('default handler', (WidgetTester tester) async {
@@ -509,7 +535,7 @@ void main() {
       await tester.pumpAndSettle();
       expect(await helper.tapText(tester, 'Tap me'), equals(1));
 
-      expect(fwfh_url_launcher.mockGetLaunchUrls(), equals(const [href]));
+      expect(mockLaunchedUrls, equals(const [href]));
     });
   });
 
@@ -555,7 +581,7 @@ void main() {
       final controller = ScrollController();
       final renderMode = ListViewMode(controller: controller);
       final html =
-          '${'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ' * 1000}'
+          '${'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ' * 999}'
           '<a name="bottom">Bottom></a>';
       final key = GlobalKey<HtmlWidgetState>();
       await explain(tester, renderMode, html: html, key: key);
