@@ -52,7 +52,8 @@ Future<void> main() async {
         equals(
           '[CssSizing:$sizingConstraints,child='
           '[SvgPicture:'
-          'pictureProvider=ExactAssetPicture(name: "$assetName", bundle: null, colorFilter: null)'
+          'pictureProvider=ExactAssetPicture(name: "$assetName", '
+          'bundle: null, colorFilter: null)'
           ']]',
         ),
       );
@@ -193,7 +194,13 @@ Future<void> main() async {
     });
   });
 
-  final goldenSkip = Platform.isLinux ? null : 'Linux only';
+  final goldenSkipEnvVar = Platform.environment['GOLDEN_SKIP'];
+  final goldenSkip = goldenSkipEnvVar == ''
+      ? Platform.isLinux
+          ? null
+          : 'Linux only'
+      : 'GOLDEN_SKIP=$goldenSkipEnvVar';
+
   GoldenToolkit.runWithConfiguration(
     () {
       group(
@@ -293,23 +300,11 @@ class _Golden extends StatelessWidget {
   Widget build(BuildContext _) => Scaffold(
         body: Padding(
           padding: const EdgeInsets.all(8.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Text(
-                contents.replaceAll(
-                  RegExp('file://.+/fwfh_svg/'),
-                  'file://.../fwfh_svg/',
-                ),
-              ),
-              const Divider(),
-              HtmlWidget(
-                contents,
-                factoryBuilder: allowDrawingOutsideViewBox
-                    ? () => _GoldenAllowFactory()
-                    : () => _GoldenDisallowFactory(),
-              ),
-            ],
+          child: HtmlWidget(
+            contents,
+            factoryBuilder: allowDrawingOutsideViewBox
+                ? () => _GoldenAllowFactory()
+                : () => _GoldenDisallowFactory(),
           ),
         ),
       );
@@ -361,11 +356,12 @@ HttpClient _createMockSvgImageHttpClient() {
     final onData =
         invocation.positionalArguments[0] as void Function(List<int>);
     final onDone =
-        invocation.namedArguments[const Symbol("onDone")] as Function?;
+        invocation.namedArguments[const Symbol('onDone')] as Function?;
     return Stream.fromIterable(<List<int>>[redTriangleBytes])
         .listen((data) async {
       await Future.delayed(const Duration(milliseconds: 10));
       onData(data);
+      // ignore: avoid_dynamic_calls
       onDone?.call();
     });
   });
