@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
@@ -7,6 +8,8 @@ import 'package:webview_flutter/webview_flutter.dart' as lib;
 import 'web_view.dart';
 
 class WebViewState extends State<WebView> {
+  final _timers = <Timer>[];
+
   late double _aspectRatio;
   String? _firstFinishedUrl;
   _Issue37? _issue37;
@@ -19,7 +22,7 @@ class WebViewState extends State<WebView> {
 
     if (widget.unsupportedWorkaroundForIssue37) {
       _issue37 = _Issue37(this);
-      WidgetsBinding.instance?.addObserver(_issue37!);
+      _widgetsBindingInstance?.addObserver(_issue37!);
     }
   }
 
@@ -66,20 +69,26 @@ class WebViewState extends State<WebView> {
 
   @override
   void dispose() {
+    for (final timer in _timers) {
+      timer.cancel();
+    }
+
     if (_issue37 != null) {
-      WidgetsBinding.instance?.removeObserver(_issue37!);
+      _widgetsBindingInstance?.removeObserver(_issue37!);
     }
 
     super.dispose();
   }
 
-  Future<String> eval(String js) =>
-      _wvc?.runJavascriptReturningResult(js).catchError((_) => '') ??
-      Future.value('');
+  Future<String> eval(String js) async {
+    try {
+      return await _wvc!.runJavascriptReturningResult(js);
+    } catch (_) {
+      return '';
+    }
+  }
 
   Future<void> _autoResize() async {
-    // TODO: enable codecov when `flutter drive --coverage` is available
-    // https://github.com/flutter/flutter/issues/7474
     if (!mounted) {
       return;
     }
@@ -145,12 +154,15 @@ class WebViewState extends State<WebView> {
           _autoResize();
         } else {
           // or wait for the specified duration
-          Future.delayed(interval).then((_) => _autoResize());
+          _timers.add(Timer(interval, _autoResize));
         }
       }
     }
   }
 }
+
+// TODO: remove workaround when our minimum Flutter version >2.12
+WidgetsBinding? get _widgetsBindingInstance => WidgetsBinding.instance;
 
 class _Issue37 with WidgetsBindingObserver {
   final WebViewState wvs;
